feat(electionmodal): added content for independent candidates
